# test: use esbuild to run