--- a/package.json
+++ b/package.json
@@ -6,7 +6,7 @@
 	"module": "dist/module/index.js",
 	"types": "dist/types/index.d.ts",
 	"scripts": {
-		"test": "TS_IGNORE_ERROR=true mocha -r ts-node/register --extension ts,js src/*.test.ts",
+		"test": "mocha -r @esbuild-kit/cjs-loader --extension ts,js src/*.test.ts",
 		"build": "rimraf dist/ && tsc && tsc -p tsconfig.module.json && tsc -p tsconfig.esm.json && node tools/post-build.js",
 		"prepublishOnly": "npm run build"
 	},
@@ -17,13 +17,13 @@
 	},
 	"license": "MIT",
 	"devDependencies": {
+		"@esbuild-kit/cjs-loader": "^2.4.1",
 		"@types/mocha": "^8.2.1",
 		"@types/node": "^14.14.35",
 		"mocha": "^8.3.2",
 		"node-pty": "^0.10.0",
 		"prettier": "^2.2.1",
 		"rimraf": "^3.0.2",
-		"ts-node": "^9.1.1",
 		"typescript": "^4.2.3"
 	},
 	"exports": {