Tests should be much faster to run:

#### Before
```
$ hyperfine 'npm run test'
Benchmark 1: npm run test
  Time (mean ± σ):      6.346 s ±  0.209 s    [User: 10.274 s, System: 0.794 s]
  Range (min … max):    6.040 s …  6.737 s    10 runs
```

#### After
```
$ hyperfine 'npm run test'
Benchmark 1: npm run test
  Time (mean ± σ):      1.107 s ±  0.061 s    [User: 0.965 s, System: 0.228 s]
  Range (min … max):    1.051 s …  1.220 s    10 runs
```